starting user home page in index.tsx
